feat(app): add APPSTORE conditional compilation for Claude CLI
What: Create two build variants — Homebrew (with Claude CLI) and
App Store (sandbox-safe, OpenAI API only).

- Extract all CLI code into ClaudeCLIProtocolGenerator.swift wrapped
  in a single top-level #if !APPSTORE
- Clean ProtocolGenerator.swift to shared-only code (prompts, file I/O)
- Guard .claudeCLI enum case, claudeBin property, default provider
  in AppSettings.swift
- Guard .claudeCLI switch case in MeetingTranscriberApp.swift
- Guard claudeBinaries state and CLI picker in SettingsView.swift
- ProtocolError retains #if !APPSTORE for CLI error cases (enum cases
  cannot be added via extension)

CaseIterable synthesizes allCases from compiled cases only, so the
provider Picker automatically excludes claudeCLI in APPSTORE builds.
UserDefaults migration: "claudeCLI" rawValue → nil → fallback to
.openAICompatible — no migration code needed.
